feat(encryption): :sparkles:  Exposed encryption driver; Cleaned up tests
